// src/api/generated/models/memberDetailResDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Fittner API
 * Fittner API with Spring Boot
 * OpenAPI spec version: 1.0.0
 */
import type { MemberDetailResDtoTicketCode } from './memberDetailResDtoTicketCode';

/**
 * 응답 데이터
 */
export interface MemberDetailResDto {
  /** 결석 카운트 */
  absenceCnt?: number;
  /** 출석 카운트 */
  attendanceCnt?: number;
  /** 회원키값 */
  memberId?: string;
  /** 회원명 */
  memberName?: string;
  /** 상품 개월수 */
  productMonth?: number;
  /** 상품명 */
  productName?: string;
  /** 회원의 상품 잔여횟수 */
  productRemainCnt?: number;
  /** 상품 총횟수 */
  productTotalCnt?: number;
  /** 티켓 상태 */
  ticketCode?: MemberDetailResDtoTicketCode;
  /** 회원의 티켓 만료일 */
  ticketEndDate?: string;
  /** 티켓키값 */
  ticketId?: string;
  /** 회원의 티켓 시작일 */
  ticketStartDate?: string;
}